+++ src/components/TaskModal/index.js
@@ -14,6 +14,7 @@ import {
 
 import Colors from './../../theme/colors';
 import DatePicker from 'react-native-datepicker';
+import styles, {datePickerStyles} from './DatePickerStyles.js';
 export default class TaskModal extends Component{
   constructor(props){
     super(props)
@@ -33,13 +34,11 @@ export default class TaskModal extends Component{
   }
 
   render() {
-
     return (
      <View style={styles.modalContainer}>
         <View style={styles.wrapperContainer}>
           <ImageBackground source={require('./../../images/pattern.png')}  style={styles.headerContainer}>
-            <Text style={styles.headerText}>   AGREGAR TAREA</Text >
-
+            <Text style={styles.headerText}>AGREGAR TAREA</Text >
               </ImageBackground>
           <View style={styles.bodyContainer}>
               <Text style={styles.addTaskText}></Text>
@@ -47,7 +46,6 @@ export default class TaskModal extends Component{
                 style={styles.inputText}
                 placeholder="Agregar Tarea"
                 onChangeText={ (text) => this.setState({taskText: text}) }/>
-                
                <DatePicker
                 style={{width: 300}}
                 date={this.state.date}
@@ -57,20 +55,9 @@ export default class TaskModal extends Component{
                 confirmBtnText="Confirm"
                 cancelBtnText="Cancel"
                 showIcon={false}
-                customStyles={{
-                  dateIcon: {
-                    position: 'center',
-                    left: 0,
-                    top: 4,
-                    marginLeft: 0
-                  },
-                  dateInput: {
-                    borderColor:'transparent',
-
-
-                    marginLeft: 0
-                  }
-                }}
+                customStyles={
+                  datePickerStyles
+                }
                 onDateChange={(date) => {this.setState({date: date})}}
               />
               <TouchableHighlight
@@ -93,87 +80,3 @@ export default class TaskModal extends Component{
     );
   }
 }
-
-
-
-// TODO: Create Theme styles (Buttons, Global Components, etc) so it can be imported on each component as required.
-const styles = StyleSheet.create({
-  modalContainer: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: 'rgba(0,0,0,.5)'
-  },
-  wrapperContainer: {
-    backgroundColor: 'white',
-    height: 567,
-  },
-  headerContainer:{
-    width:'100%',
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: Colors.primaryColorDarker
-
-  },
-  headerText:{
-    width: 300,
-    fontFamily: 'Roboto',
-    color: Colors.white,
-    fontSize: 30,
-    textAlign: 'center'
-   
-  },
-  bodyContainer: {
-    flex: 1,
-    height: 315,
-    justifyContent: 'space-around',
-    alignItems: 'center',
-    padding: 20,
-  },
-  addTaskText: {
-    fontFamily: 'Roboto',
-    fontSize: 16,
-    textAlign: 'center'
-  },
-  inputText:{
-    width: 300,
-    fontFamily: 'Roboto',
-    fontSize: 30,
-    textAlign: 'center',
-    borderBottomWidth: Platform.OS === 'ios' ? 2 : 0,
-    borderBottomColor: Colors.separator
-  },
-  addButton: {
-      width: 325,
-      height: 55,
-      paddingTop: 15,
-      borderRadius: 27,
-      borderStyle: 'solid',
-      borderWidth: 2,
-      backgroundColor: Colors.primaryColorDarker,
-      borderColor: Colors.primaryColorDarker
-  },
-  addButtonText: {
-    fontFamily: 'Roboto',
-    textAlign: 'center',
-    fontSize: 17,
-    color: 'white'
-  },
-  cancelButton: {
-      width: 325,
-      height: 55,
-      paddingTop: 15,
-      borderRadius: 27,
-      borderStyle: 'solid',
-      borderWidth: 2,
-      backgroundColor: 'white',
-      borderColor: Colors.primaryColorDarker
-  },
-  cancelButtonText: {
-    fontFamily: 'Roboto',
-    textAlign: 'center',
-    fontSize: 17,
-    color: Colors.primaryColorDarker
-  }
-});
